backend/requirements.txt
django
channels
channels-redis
daphne
solana
solders
base58
djangorestframework
